* fixup to InitValsIncrement and others -- no longer depends on !is_matrix and also updated DataTable version * taMath vec_convolve now uses frame index view -- safe for filtered data tables * put all the TwoD layer stuff into base LeabraLayer and removed subtype -- much easier for switching back and forth and avoiding many problems with changing type of layer * Leabra layer average activation updating: added "use_first" option (default TRUE -- will change learning behavior of existing models, most likely for the better) -- uses the first real sample instead of our likely incorrect guess from targ_init..

+++ src/temt/ta_data/DataCol.cpp
@@ -819,7 +819,7 @@ bool DataCol::InitValsByIncrement(float first_value, float increment, int st_row
   }
   else {
     for(int i=0; i<n_rows; i++) {
-      SetValAsFloat(first_value + i*increment, st_row + i);
+      SetValAsFloat(first_value + (float)i*increment, st_row + i);
     }
   }
   return true;
